+++ kafkametrics/datadog/datadog_test.go
@@ -6,8 +6,6 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/DataDog/kafka-kit/kafkametrics"
-
 	dd "github.com/zorkian/go-datadog-api"
 )
 
@@ -70,71 +68,4 @@ func mockSeriesWithoutPoints() []dd.Series {
 	}
 
 	return ss
-}
-
-// This is essentially tested via TestGetHostTagMap
-// and TestPopulateFromTagMap.
-// func TestBrokerMetricsFromList(t *testing.T) {}
-
-// func TestGetHostTagMap(t *testing.T) {}
-
-func TestPopulateFromTagMap(t *testing.T) {
-	b := kafkametrics.BrokerMetrics{}
-
-	// Test with complete input.
-	tagMap := mockTagMap()
-	err := populateFromTagMap(b, map[string][]string{}, tagMap, "broker_id")
-	if err != nil {
-		t.Errorf("Unexpected error: %s\n", err)
-	}
-
-	// Keep a broker reference
-	// for the next test.
-	var rndBroker *kafkametrics.Broker
-
-	for id, broker := range b {
-		rndBroker = broker
-		if broker.ID != id {
-			t.Errorf("Expected ID %d, got %d\n", id, broker.ID)
-		}
-		if broker.InstanceType != "mock" {
-			t.Errorf("Expected broker InstanceType mock, got %s\n", broker.InstanceType)
-		}
-	}
-
-	// Test with incomplete input.
-	tagMap[rndBroker] = tagMap[rndBroker][1:]
-	err = populateFromTagMap(b, map[string][]string{}, tagMap, "broker_id")
-	if err == nil {
-		t.Errorf("Expected error, got nil")
-	}
-}
-
-func mockTagMap() map[*kafkametrics.Broker][]string {
-	tm := map[*kafkametrics.Broker][]string{}
-
-	for i := 0; i < 5; i++ {
-		bid := 1000 + i
-		b := &kafkametrics.Broker{
-			ID:           bid,
-			Host:         fmt.Sprintf("host%d", i),
-			InstanceType: "mock",
-			NetTX:        1073741824.00,
-		}
-
-		bidTag := fmt.Sprintf("broker_id:%d", bid)
-		tm[b] = []string{bidTag, "instance-type:mock"}
-	}
-
-	return tm
-}
-
-// This tests both tagValFromScope and valFromTags.
-func TestTagValFromScope(t *testing.T) {
-	series := mockSeries()
-	v := tagValFromScope(series[0].GetScope(), "instance-type")
-
-	if v != "mock" {
-		t.Errorf("Expected tag val mock, got %s\n", v)
-	}
 }
